Initial memory manager to support paging is in place
The page directory and the page table containing the last page are both permanently mapped
at a predictable address, because the last page is used to map page tables in and out of
physical memory so they can be edited to make mappings. Essentially, when a mapping is made
for vaddr, the page table of that vaddr is mapped at 0xFFFFF000 (which can be done because
the page directory and the last page table are mapped in) and the mapping is then made.